Extract SettingsPanel to layout for global access
- Create SettingsPanel.svelte component that reads state from store
- Render SettingsPanel in layout so it's accessible from any page
- Remove settings panel code from OTPList.svelte
- Simplify Header to just toggle settings visibility
- Remove unused props from OTPList and +page.svelte

Now clicking the settings gear icon from any page will show/hide
the settings panel without navigation.
